Organise some CLI graph construction stuff.
